+++ projects/ccf-eui/src/environments/environment.ts
@@ -15,26 +15,19 @@ export const environment = {
   dbOptions: {
     ccfOwlUrl: 'assets/ccf.owl.n3store.json',
     ccfContextUrl: 'https://hubmapconsortium.github.io/ccf-ontology/ccf-context.jsonld',
-    // dataSources: [
-    //   'https://hubmapconsortium.github.io/hra-registrations/federated/rui_locations.jsonld',
-    //   'https://apps.humanatlas.io/hra-api/v1/gtex/rui_locations.jsonld'
-    // ],
     hubmapDataService: 'search-api',
     hubmapDataUrl: 'https://search.api.hubmapconsortium.org/v3/entities/search',
     hubmapToken: localStorage.getItem('HUBMAP_TOKEN') ?? '',
 
     useRemoteApi: false,
     remoteApiEndpoint: 'https://apps.humanatlas.io/hra-api--staging/v1'
-    // remoteApiEndpoint: 'https://apps.humanatlas.io/hra-api/v1'
-    // remoteApiEndpoint: 'http://localhost:8080/v1'
   },
   customization: {
     theme: 'default',
     header: true,
     homeUrl: 'https://portal.hubmapconsortium.org/',
     logoTooltip: 'Human BioMolecular Atlas Project',
-    loginDisabled: false,
-    // filter: { sex: 'Male' }
+    loginDisabled: false
   },
   googleAnalyticsToken: 'G-B3DT7XPMRT'
 };
